refactor: add Service function from SignIn and SignOn
