+++ src/components/Signin/signin.ts
@@ -2,6 +2,11 @@ import Cookies from "js-cookie";
 import { IToken, TLoginUser } from "../../types/api.interface";
 import { loginUser } from "../Api/register-login.api";
 import { Main } from "../Main/Main";
+import {
+  checkRequired,
+  checkEmail,
+  checkPassword,
+} from "../Service/auth.login.service";
 
 export function SignIn(): void {
   const main = document.querySelector("main") as HTMLElement;
@@ -39,65 +44,6 @@ function FormSignIn(): void {
   const form = document.getElementById("form") as HTMLFormElement;
   const email = document.getElementById("email") as HTMLInputElement;
   const password = document.getElementById("password") as HTMLInputElement;
-
-  const eml =
-    // eslint-disable-next-line no-useless-escape
-    /^(([^<>()\[\]\\.,;:\s@"]+(\.[^<>()\[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
-
-  const Validate = (
-    item: HTMLElement,
-    messageType: "success" | "error",
-    message?: string
-  ) => {
-    const formControl = item.parentElement as HTMLDivElement;
-    formControl.className =
-      messageType === "error" ? "form-control error" : "form-control success";
-    if (messageType === "error" && !!message) {
-      const small = formControl.querySelector("small");
-      (small as HTMLInputElement).innerHTML = message;
-    }
-  };
-
-  function checkEmail(email: HTMLInputElement) {
-    if (eml.test(email.value.toLowerCase())) {
-      Validate(email, "success");
-      return true;
-    } else {
-      Validate(email, "error", "Example: email@email.en");
-      return false;
-    }
-  }
-
-  function checkPassword(password: HTMLInputElement) {
-    if (password.value.toLowerCase().length >= 6) {
-      Validate(password, "success");
-      return true;
-    } else {
-      Validate(password, "error", "Password length min 6 symbols");
-      return false;
-    }
-  }
-
-  const checkRequired = (items: HTMLInputElement[]) => {
-    let count = 0;
-    items.forEach((item: HTMLInputElement) => {
-      if (item.value.trim() === "") {
-        Validate(item, "error", captializedNameOFInput(item) + " is required");
-      } else {
-        Validate(item, "success");
-        count += 1;
-        if (+count === 2) {
-          /**-----------Aleksey form export to server---------*/
-          setTimeout(() => console.log("Validate Ok"), 2000);
-          /**-----------Aleksey form export to server---------*/
-        }
-      }
-    });
-  };
-
-  const captializedNameOFInput = (item: HTMLInputElement) => {
-    return item.id[0].toUpperCase() + item.id.slice(1);
-  };
 
   form.addEventListener("submit", async function (e: Event) {
     let body: TLoginUser = { email: "", password: "" };
